Allowing on-site override of default logged-in/anon icons, cleanup
This commit allows the default logged-in/anon icons to be overriden through two system messages (refreshed-icon-logged-in and refreshed-icon-logged-out). If the appropriate message for the current situation is not empty (and SocialProfile is not installed), the message's contents are placed in the src of the avatar img element, overriding the default WikiFont logged-in/anon icons. The avatar logic has also been cleaned up to reduce redundancy. Finally, the avatar img element no longer has a hard-coded width attribute as this is already handled with CSS (which allows for more on-site customization).

Bug: T163961
Change-Id: I7851c5f63282b951c6b0b5baa59f84f8fa088d48