Move indicators to be consistent with other skins
Moves the indicators above firstHeading to be consistent with
other skins such as Vector. CSS is changed as appropriate. This
improves the indicators' behavior when they or firstHeading are
very large. Instead of rendering behind firstHeading as was the
case before, now the indicators force firstHeading to wrap to
multiple lines.

Change-Id: I551ce2c3ae5447297333f1c505a4f5bfcf250dda