Use OOUI icons and new dropdown system, refactor PHP, etc.
- Began switch from WikiFont to OOUI icons (inline SVGs) on
large and medium. Added icons for some tools that didn't have
icons before. Also, (finally!) switched the hamburger icon to
a vector instead of a PNG.
- Implemented a new dropdown system in the header. It does not
use JS and opens on hover instead of click. (In the future we may
want to add JS to change dropdowns' aria attributes when they're
expanded.)
- Redid the header to use flexbox for ease of use. (Looking forward,
is flexbox a good way to finally center images/svgs in the header?)
- Began notable PHP refactoring. The goal is to make
RefreshedTemplate.php more readable by handling logic in small,
well-documented functions rather than having lots of logic alongside
lots of markup in execute().

Note the new dropdown and header markup is different than it was, so
any wikis with custom CSS for Refreshed 3 will need to redo it when
Refreshed 4 releases.

Change-Id: I853b3e4f4614a7a0584e9a1bb356a9d9521a7e2d