Add explore button, rtl icons, minor CSS changes, minor i18n
- Fix padding of icon-only buttons on small
- Fix z-index of Echo dropdown so it appears in front of header
- Change related changes icon
- Add cite this page, next icons
- Add RTL icons
- Add explore button. Explore button pulls "Explore" message from
MediaWiki:Refreshed-explore. When JS is disabled or the new global
variable $wgRefreshedUseExploreWithoutHeaderCategories is true (default
is false), the explore button appears on medium and big at all times,
and the header category buttons never appear. When JS is enabled or
the variable is false, the explore button only appears when the
window is too narrow to fit all the header categories.
- i18n

Change-Id: Ib61bddd340252e58a3cb0ceed47fede24bba0b35