Fix issue with UTF-8 characters in content action menus being garbled
Simplest possible test case: load any page with ?uselang=fi (or any other UTF-8 language of your choosing).
Observe how translations for things like the "View" tab are garbled (i.e. for Finnish the translation of that tab is "Näytä" but prior to this patch it'd render as "NÃ¤ytÃ¤".

Change-Id: Iaa51293cc43e79d86372a78e8508b750d82b87fd