Add support for i18n messages to sidebar headers to correctly translate default sidebar and render extension-added blocks
For Extension:NewsBox, previously a literal section named
"newsbox-title" would show up in the sidebar instead of the parsed
value of said message (e.g. "ShoutWiki messages").
This fixes that issue and now the correct value is shown in the sidebar
instead of the message key.

On unconfigured MediaWiki installs, the 'navigation' header would always
show 'navigation' regardless of language.

Change-Id: I3ce1ed04ec10e93fd144d17e44ccb040bf207421